did some file cleanup and finished implementing jsonwebtoken
