+++ lib/middleware/auth.js
@@ -1,6 +1,6 @@
 const { decode } = require('js-base64');
 
-const { userModel } = require('../models/model');
+const { userModel } = require('../models');
 
 const decodeBase64 = (encoded) => {
   const data = decode(encoded).toString().split(':');
