+++ lib/models/index.js
@@ -2,7 +2,7 @@
 
 const userSchema = require('./users-schema');
 const charSchema = require('./character-schema');
-const campaignSchema = require('./campaign-schema')
+const campaignSchema = require('./campaign-schema');
 
 class Model {
   constructor(schema) {
@@ -13,8 +13,7 @@ class Model {
     try {
       let recordToAdd = new this.schema(record);
       return await recordToAdd.save();
-    }
-    catch (e) {
+    } catch (e) {
       console.error('ERROR CREATING RECORD', e);
       return false;
     }
@@ -24,8 +23,7 @@ class Model {
     try {
       let record = await this.schema.findOne({ _id });
       return record;
-    }
-    catch (e) {
+    } catch (e) {
       console.error('ERROR FINDING RECORD', e);
       return false;
     }
@@ -35,8 +33,7 @@ class Model {
     try {
       let records = await this.schema.find(query);
       return records;
-    }
-    catch (e) {
+    } catch (e) {
       console.error('ERROR FINDING RECORD', e);
       return false;
     }
@@ -44,11 +41,13 @@ class Model {
 
   async update(_id, changedRecord) {
     try {
-      let newRecord = await this.schema.findByIdAndUpdate({ _id }, changedRecord);
+      let newRecord = await this.schema.findByIdAndUpdate(
+        { _id },
+        changedRecord,
+      );
       console.log('Updated Record:', newRecord);
       return newRecord;
-    }
-    catch (e) {
+    } catch (e) {
       console.error('ERROR UPDATING RECORD', e);
       return false;
     }
@@ -58,8 +57,7 @@ class Model {
     try {
       let deletedRecord = await this.schema.deleteOne({ _id });
       return _id;
-    }
-    catch (e) {
+    } catch (e) {
       console.error('ERROR DELETING RECORD', e);
       return false;
     }
@@ -70,4 +68,4 @@ module.exports = {
   userModel: new Model(userSchema),
   charModel: new Model(charSchema),
   campaignModel: new Model(campaignSchema),
-}
+};
