+++ lib/server.js
@@ -3,7 +3,7 @@
 const express = require('express');
 const cors = require('cors');
 const morgan = require('morgan');
-const router = require('./routes');
+const { campaignRouter, characterRouter, userRouter } = require('./routes');
 const { swagger, options } = require('../docs/swagger');
 const errorHandler = require('./middleware/error-handler.js');
 
@@ -11,37 +11,30 @@ const app = express();
 const server = require('http').createServer(app);
 const io = require('socket.io')(server);
 
-io.on('connection', socket => {
+io.on('connection', (socket) => {
   console.log('Socket connected:', socket.id);
 
-  socket.on('join', payload => {
+  socket.on('join', (payload) => {
     socket.join(payload);
     console.log('Socket joined room:', payload);
   });
 
-  socket.on('test', payload => {
+  socket.on('test', (payload) => {
     console.log('test');
-    io
-      .to(payload.room)
-      .emit('hello', payload);
+    io.to(payload.room).emit('hello', payload);
   });
 
-  socket.on('roll', payload => {
+  socket.on('roll', (payload) => {
     console.log('server roll');
 
-    io
-      .to(payload.room)
-      .emit('roll', payload);
+    io.to(payload.room).emit('roll', payload);
   });
 
-  socket.on('character-join', payload => {
+  socket.on('character-join', (payload) => {
     console.log('room:', payload.room);
     console.log('character:', payload.character);
-    io
-      .to(payload.room)
-      .emit('character-joined', payload.character);
+    io.to(payload.room).emit('character-joined', payload.character);
   });
-  
 });
 
 const expressSwagger = swagger(app);
@@ -62,7 +55,9 @@ app.get('/', (req, res) => {
 });
 
 // API v1
-app.use('/v1/api', router);
+app.use('/v1/api', campaignRouter);
+app.use('/v1/api', userRouter);
+app.use('/v1/api', characterRouter);
 
 app.use(errorHandler);
